Api 주소 겹ì¹치는 오류 수정
